* native/jni/java-nio/javanio.c (cpnio_fcntl): Last argument is long.         * native/jni/java-nio/javanio.h (cpnio_fcntl): Likewise.         * native/jni/java-nio/gnu_java_nio_VMChannel.c         (Java_gnu_java_nio_VMChannel_lock): Likewise.         (Java_gnu_java_nio_VMChannel_unlock): Likewise.

+++ native/jni/java-nio/javanio.c
@@ -104,7 +104,7 @@ cpnio_recvfrom (int fd, void *buf, size_t len, int flags,
 }
 
 extern inline int
-cpnio_fcntl (int fd, int cmd, int arg)
+cpnio_fcntl (int fd, int cmd, long arg)
 {
 #ifdef HAVE_FCNTL
   return fcntl (fd, cmd, arg);
